Removed previous fix for initialisation, this was not a fix. The problem was actually that the mongo database was not configured correctly, and the router would not correctly initialise.
I've tried to improve the error messaging in this case.

I've also modified the default read operation of the router so that it always reads from master.
I've done this as we're not particularly worried about performance at this point, and it simplifies
some of the incosistency issues that people have been having where the router would report that a route
didn't exist if it was queried immediately after creation due to the mongo replication not having taken place.